+++ khoros/utils/tests/test_messages.py
@@ -4,7 +4,7 @@
 :Synopsis:       This module is used by pytest to verify that messages function properly
 :Created By:     Jeff Shurtliff
 :Last Modified:  Jeff Shurtliff
-:Modified Date:  11 Mar 2021
+:Modified Date:  26 Sep 2021
 """
 
 import pytest
@@ -132,6 +132,47 @@ def test_construct_with_tag_iterables():
             assert payload == control_data      # nosec
         except AssertionError:
             assert_tags_present(payload, ['hello', 'world'])
+    return
+
+
+def test_payload_validation():
+    """This function tests the validation of the message payload to ensure invalid data raises an exception.
+
+    .. versionadded:: 4.3.0
+    """
+    # Test null payload
+    with pytest.raises(exceptions.InvalidMessagePayloadError):
+        messages.validate_message_payload(payload=None)
+
+    # Test conversion to dictionary when JSON string
+    payload = '{"data": {"type": "message", "subject": "This is a message subject"}}'
+    payload = messages.validate_message_payload(payload)
+    assert isinstance(payload, dict)
+
+    # Test incorrect data type
+    with pytest.raises(exceptions.InvalidMessagePayloadError):
+        payload = [{'type': 'message'}, {'subject': 'This is a message subject'}]
+        messages.validate_message_payload(payload)
+
+    # Test payload that is not wrapped in the 'data' field
+    with pytest.raises(exceptions.InvalidMessagePayloadError):
+        payload = {'type': 'message', 'subject': 'This is a message subject'}
+        messages.validate_message_payload(payload)
+
+    # Test payload that is missing the 'type' sub-field
+    with pytest.raises(exceptions.InvalidMessagePayloadError):
+        payload = {'subject': 'This is a message subject'}
+        messages.validate_message_payload(payload)
+
+    # Test payload that has the incorrect 'type' value
+    with pytest.raises(exceptions.InvalidMessagePayloadError):
+        payload = {'type': 'article', 'subject': 'This is a message subject'}
+        messages.validate_message_payload(payload)
+
+    # Test valid payload
+    payload = {'data': {'type': 'message', 'subject': 'This is a message subject'}}
+    payload = messages.validate_message_payload(payload)
+    assert payload.get('data').get('type') == 'message'
     return